+++ src/assets/themes/GlobalStyle.ts
@@ -3,22 +3,25 @@ import { createGlobalStyle } from "styled-components";
 const GlobalStyle = createGlobalStyle`
   *, *::before, *::after {
     margin: 0;
-    padding: 0;
     box-sizing: border-box;
   }
 
   html, body {
-    color: ${props => props.theme.colors.text.dark};
-    background-color: ${props => props.theme.colors.accent_white_0};
-    background-image: linear-gradient(to bottom left, ${props => props.theme.colors.accent_white_0}, ${props => props.theme.colors.accent_white_1});
+    ${({ theme }) => `
+      color: ${theme.colors.text.dark};
+      background-color: ${theme.colors.accent_white_0};
+      background-image: linear-gradient(to bottom left, ${theme.colors.accent_white_0}, ${theme.colors.accent_white_1});
 
-    height: 100dvh;
+      font-family: ${theme.font.family.default};
+      scroll-padding: ${theme.spacing.s.rem};
+    `}
 
-    font-family: ${props => props.theme.font.family.default};
+    min-height: 100dvh;
+    scroll-behavior: smooth;
     overscroll-behavior: none;
   }
   #root {
-    height: inherit;
+    min-height: inherit;
     overscroll-behavior: none;
   }
 `;