+++ src/contexts/ApiContext/ApiContext.tsx
@@ -5,25 +5,25 @@ const config: ApiContextType = {
   baseUrl: 'https://api.quotly.eu',
   routes: {
     quotes: {
-      path: '/v1/quotes',
-      subPaths: {
-        comments: '/comments',
-        reactions: '/reactions',
+      construct: (id?) => `${config.baseUrl}/v1/quotes${id ? `/${id}` : ''}`,
+      sub: {
+        comments: (id) => `${config.routes.quotes.construct(id)}/comments`,
+        reactions: (id) => `${config.routes.quotes.construct(id)}/reactions`,
       }
     },
     roles: {
-      path: '/v1/roles'
+      construct: (roleId?) => `${config.baseUrl}/v1/roles${roleId ? `/${roleId}` : ''}`,
     },
     users: {
-      path: '/v1/users',
-      subPaths: {
-        reactions: '/reactions',
-        roles: '/roles',
-        savedQuotes: '/saved-quotes'
+      construct: (discordId?) => `${config.baseUrl}/v1/users${discordId ? `/${discordId}` : ''}`,
+      sub: {
+        reactions: (discordId) => `${config.routes.users.construct(discordId)}/reactions`,
+        roles: (discordId) => `${config.routes.users.construct(discordId)}/roles`,
+        savedQuotes: (discordId) => `${config.routes.users.construct(discordId)}/saved-quotes`,
       }
     }
   },
-  discordAuth: 'https://discord.com/oauth2/authorize?client_id=1303517823452184697&response_type=code&redirect_uri=https%3A%2F%2Fquotly.eu%2Foauth&scope=identify+email' 
+  discordAuth: 'https://discord.com/oauth2/authorize?client_id=1303517823452184697&response_type=code&redirect_uri=http%3A%2F%2Flocalhost%3A3570%2Foauth&scope=identify+email' 
 };
 
 /**