+++ src/pages/_MarkdownPage/MarkdownPage.type.ts
@@ -1,0 +1,12 @@
+import { Options } from "react-markdown";
+
+/**
+ * MarkdownPageType type for MarkdownPage
+ * @example
+ * <MarkdownPage children=''
+ */
+export type MarkdownPageType = Options & {
+  childrenPre?: React.ReactNode;
+  childrenSuf?: MarkdownPageType['childrenPre'];
+  title?: string;
+};